requirements.txt
Flask>=3.0,<4
Flask-Login>=0.6
Flask-SQLAlchemy>=3.1
Flask-WTF>=1.2
WTForms>=3.1
gunicorn>=21.2

